Change padding and margin to padding plus border
Padding forces overflow on a page that goes too small. Margin does not
add background color. Border does not have either problem.

Using a mix of padding and border in place of margin with padding or
padding on it's own solves the issue of background color overflow which
occurs when the page shrinks